last changes

+++ src/matrix.cpp
@@ -6,19 +6,19 @@ matrix::matrix()
     data= new double*;
     data[0]= new double;
 }
-matrix::matrix(int n)
+matrix::matrix(uint8_t n)
 {
     this->lines=this->columns=n;
     data= new double*[n];
-    for(int i=0; i<n; i++)
+    for(uint8_t i=0; i<n; i++)
         data[i]= new double[n];
 }
-matrix::matrix(int lines, int columns)
+matrix::matrix(uint8_t lines, uint8_t columns)
 {
     this->lines=lines;
     this->columns=columns;
     data= new double*[lines];
-    for(int i=0; i<lines; i++)
+    for(uint8_t i=0; i<lines; i++)
         data[i]= new double[columns];
 }
 matrix::matrix(initializer_list<initializer_list<double>> elem)
@@ -26,21 +26,21 @@ matrix::matrix(initializer_list<initializer_list<double>> elem)
     this->lines= elem.size();
     this->columns= elem.begin()->size();
     data= new double*[lines];
-    for(int i=0; i<lines; i++)
+    for(uint8_t i=0; i<lines; i++)
         data[i]= new double[columns];
     bool fimL= false, fimC= false;
-    for(int i=0; i<this->lines; i++)
+    for(uint8_t i=0; i<this->lines; i++)
     {
         if((elem.begin()+i) == elem.end())
             fimL= true;
         if(fimL)
         {
-            for(int j=0; j<this->columns; j++)
+            for(uint8_t j=0; j<this->columns; j++)
                 this->data[i][j]= 0;
         }
         else
         {
-            for(int j=0; j<this->columns; j++)
+            for(uint8_t j=0; j<this->columns; j++)
             {
                 if(((elem.begin()+i)->begin()+j) == (elem.begin()+i)->end())
                     fimC= true;
@@ -56,11 +56,11 @@ matrix::matrix(initializer_list<initializer_list<double>> elem)
 matrix::~matrix()
 {
 }
-int matrix::getLines() const
+uint8_t matrix::getLines() const
 {
     return lines;
 }
-int matrix::getColumns() const
+uint8_t matrix::getColumns() const
 {
     return columns;
 }
@@ -68,7 +68,7 @@ double** matrix::getData() const
 {
     return data;
 }
-double& matrix::at(int i, int j)
+double& matrix::at(uint8_t i, uint8_t j)
 {
     if(i >= 0 && i < this->lines && j >= 0 && j < this->columns)
         return this->data[i][j];
@@ -78,8 +78,8 @@ bool matrix::operator == (matrix x)
 {
     if(this->lines != x.lines || this->columns != x.columns)
         return false;
-    for(int i=0; i<x.lines; i++)
-        for(int j=0; j<x.columns; j++)
+    for(uint8_t i=0; i<x.lines; i++)
+        for(uint8_t j=0; j<x.columns; j++)
             if(this->data[i][j] != x.data[i][j])
                 return false;
     return true;
@@ -93,21 +93,21 @@ matrix matrix::operator =(initializer_list<initializer_list<double>> elem)
     this->lines= elem.size();
     this->columns= elem.begin()->size();
     data= new double*[lines];
-    for(int i=0; i<lines; i++)
+    for(uint8_t i=0; i<lines; i++)
         data[i]= new double[columns];
     bool fimL= false, fimC= false;
-    for(int i=0; i<this->lines; i++)
+    for(uint8_t i=0; i<this->lines; i++)
     {
         if((elem.begin()+i) == elem.end())
             fimL= true;
         if(fimL)
         {
-            for(int j=0; j<this->columns; j++)
+            for(uint8_t j=0; j<this->columns; j++)
                 this->data[i][j]= 0;
         }
         else
         {
-            for(int j=0; j<this->columns; j++)
+            for(uint8_t j=0; j<this->columns; j++)
             {
                 if(((elem.begin()+i)->begin()+j) == (elem.begin()+i)->end())
                     fimC= true;
@@ -127,10 +127,10 @@ matrix matrix::operator =(matrix x)
     lines= x.lines;
     columns= x.columns;
     data= new double*[lines];
-    for(int i=0; i<lines; i++)
+    for(uint8_t i=0; i<lines; i++)
         data[i]= new double[columns];
-    for(int i=0; i<lines; i++)
-        for(int j=0; j<columns; j++)
+    for(uint8_t i=0; i<lines; i++)
+        for(uint8_t j=0; j<columns; j++)
             data[i][j]= x.data[i][j];
     return *this;
 }
@@ -139,8 +139,8 @@ matrix matrix::operator +(matrix x)
     if(x.columns != this->columns || x.lines != this->lines)
         throw "Operacao Invalida matrix soma";
     matrix ret(this->lines, this->columns);
-    for(int i=0; i<this->lines; i++)
-        for(int j=0; j<this->columns; j++)
+    for(uint8_t i=0; i<this->lines; i++)
+        for(uint8_t j=0; j<this->columns; j++)
             ret.data[i][j]= this->data[i][j]+x.data[i][j];
     return ret;
 }
@@ -153,8 +153,8 @@ matrix matrix::operator -(matrix x)
     if(x.columns != this->columns || x.lines != this->lines)
         throw "Operacao Invalida matrix subtracao";
     matrix ret(this->lines, this->columns);
-    for(int i=0; i<this->lines; i++)
-        for(int j=0; j<this->columns; j++)
+    for(uint8_t i=0; i<this->lines; i++)
+        for(uint8_t j=0; j<this->columns; j++)
             ret.data[i][j]= this->data[i][j]-x.data[i][j];
     return ret;
 }
@@ -167,11 +167,11 @@ matrix matrix::operator *(matrix x)
     if(this->columns != x.lines)
         throw "Operacao Invalida matrix multiplicacao";
     matrix ret(this->lines, x.columns);
-    for(int i=0; i<this->lines; i++)
-        for(int j=0; j<x.columns; j++)
+    for(uint8_t i=0; i<this->lines; i++)
+        for(uint8_t j=0; j<x.columns; j++)
             {
                 ret.data[i][j]=0;
-                for(int k=0; k<this->columns; k++)
+                for(uint8_t k=0; k<this->columns; k++)
                     ret.data[i][j]+= this->data[i][k]*x.data[k][j];
             }
     return ret;
@@ -183,8 +183,8 @@ matrix matrix::operator *=(matrix x)
 matrix matrix::operator *(double x)
 {
     matrix ret(this->lines, this->columns);
-    for(int i=0; i<this->lines; i++)
-        for(int j=0; j<this->columns; j++)
+    for(uint8_t i=0; i<this->lines; i++)
+        for(uint8_t j=0; j<this->columns; j++)
             ret.data[i][j]= this->data[i][j]*x;
     return ret;
 }
@@ -197,8 +197,8 @@ matrix matrix::operator /(double x)
     if(x == 0)
         throw "Erro matrix div/0";
     matrix ret(this->lines, this->columns);
-    for(int i=0; i<this->lines; i++)
-        for(int j=0; j<this->columns; j++)
+    for(uint8_t i=0; i<this->lines; i++)
+        for(uint8_t j=0; j<this->columns; j++)
             ret.data[i][j]= this->data[i][j]/x;
     return ret;
 }
@@ -208,9 +208,9 @@ matrix matrix::operator /=(double x)
 }
 ostream& operator << (ostream& out, matrix x)
 {
-    for(int i=0; i<x.lines; i++)
+    for(uint8_t i=0; i<x.lines; i++)
     {
-        for(int j=0; j<x.columns; j++)
+        for(uint8_t j=0; j<x.columns; j++)
         {
             out << x.data[i][j] << " ";
         }
@@ -221,17 +221,17 @@ ostream& operator << (ostream& out, matrix x)
 
 matrix operator -(matrix x)
 {
-    for(int i=0; i<x.lines; i++)
-        for(int j=0; j<x.columns; j++)
+    for(uint8_t i=0; i<x.lines; i++)
+        for(uint8_t j=0; j<x.columns; j++)
             x.data[i][j]= -x.data[i][j];
     return x;
 }
 
-matrix matrix::identity(int n)
+matrix matrix::identity(uint8_t n)
 {
     matrix ret(n);
-    for(int i=0; i<n; i++)
-        for(int j=0; j<n; j++)
+    for(uint8_t i=0; i<n; i++)
+        for(uint8_t j=0; j<n; j++)
             if(i == j)
                 ret.data[i][i]= 1;
             else
@@ -241,8 +241,8 @@ matrix matrix::identity(int n)
 matrix matrix::transposta(matrix x)
 {
     matrix ret(x.columns, x.lines);
-    for(int i=0; i<x.lines; i++)
-        for(int j=0; j<x.columns; j++)
+    for(uint8_t i=0; i<x.lines; i++)
+        for(uint8_t j=0; j<x.columns; j++)
             ret.data[j][i]= x.data[i][j];
     return ret;
 }
